fixed some warnings in OSResult.cpp

+++ OS/src/OSCommonInterfaces/OSResult.cpp
@@ -7975,6 +7975,7 @@ bool GeneralResult::setRandom(double density, bool conformant)
     if (OSRand() <= density)
     {
         generalStatus = new GeneralStatus();
+
         generalStatus->setRandom(density, conformant);
     }
     if (OSRand() <= density)
@@ -8509,7 +8510,7 @@ bool VarValue::setRandom(double density, bool conformant)
 #ifdef DEBUG_OSRESULT
     cout << "Set random VarValue" << endl;
 #endif
-    this->idx = (10*OSRand());
+    this->idx = (int)(10*OSRand());
     if (OSRand() <= 0.5) this->value = 3.14156;
     else                 this->value = 2.71828;
 
@@ -8548,7 +8549,7 @@ bool VarValueString::setRandom(double density, bool conformant)
 #ifdef DEBUG_OSRESULT
     cout << "Set random VarValueString" << endl;
 #endif
-    this->idx = (10*OSRand());
+    this->idx = (int)(10*OSRand());
     if (OSRand() <= 0.5) this->value = "random string";
     else                 this->value = "";
 
@@ -8614,7 +8615,7 @@ bool OtherVarResult::setRandom(double density, bool conformant)
 #ifdef DEBUG_OSRESULT
     cout << "Set random OtherVarResult" << endl;
 #endif
-    this->idx = (10*OSRand());
+    this->idx = (int)(10*OSRand());
     if (OSRand() <= 0.5) this->value = "random string";
     else                 this->value = "";
 
@@ -8840,7 +8841,7 @@ bool DualVarValue::setRandom(double density, bool conformant)
 #ifdef DEBUG_OSRESULT
     cout << "Set random DualVarValue" << endl;
 #endif
-    this->idx = (5*OSRand());
+    this->idx = (int)(5*OSRand());
     if (OSRand() <= 0.5) this->value = 3.14156;
     else                 this->value = 2.71828;
 
@@ -8896,6 +8897,7 @@ bool OtherConstraintResult::setRandom(double density, bool conformant)
         }
     }
 
+
     return true;
 }//OtherConstraintResult::setRandom
 
@@ -8905,7 +8907,7 @@ bool OtherConResult::setRandom(double density, bool conformant)
 #ifdef DEBUG_OSRESULT
     cout << "Set random OtherConResult" << endl;
 #endif
-    this->idx = (5*OSRand());
+    this->idx = (int)(5*OSRand());
     if (OSRand() <= 0.5) this->value = "random string";
     else                 this->value = "";
